requirements.txt
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
black==22.1.0
cffi==1.15.0
click==8.0.3
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.73.0
fastapi-pagination==0.9.1
gunicorn==20.1.0
h11==0.13.0
httptools==0.3.0
idna==3.3
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pymongo==4.0.1
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.0
typing_extensions==4.0.1
uvicorn==0.17.1
uvloop==0.16.0
watchgod==0.7
websockets==10.1
